refactor: separando as entidades por contexto
