--- src/main/java/br/com/solutis/locadora/model/entity/GenderEnum.java
@@ -1,6 +1,0 @@
-package br.com.solutis.locadora.model.entity;
-
-public enum GenderEnum {
-    MALE,
-    FEMALE,
-}
