+++ src/main/java/br/com/solutis/locadora/model/entity/car/Model.java
@@ -1,5 +1,6 @@
-package br.com.solutis.locadora.model.entity;
+package br.com.solutis.locadora.model.entity.car;
 
+import br.com.solutis.locadora.model.entity.AbstractEntity;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -14,19 +15,19 @@ import java.util.List;
 @NoArgsConstructor
 @ToString
 @SuperBuilder
-@Table(name = "car_models")
-public class CarModel extends AbstractEntity {
+@Table(name = "models")
+public class Model extends AbstractEntity {
     @Column(nullable = false)
     private String description;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "category", nullable = false)
-    private CarModelCategoryEnum category;
+    private ModelCategoryEnum category;
 
     @JsonIgnoreProperties("models")
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "manufacturer_id", nullable = false)
-    private CarManufacturer manufacturer;
+    private Manufacturer manufacturer;
 
     @JsonIgnoreProperties("models")
     @OneToMany(mappedBy = "model", fetch = FetchType.LAZY)
