+++ src/main/java/br/com/solutis/locadora/model/entity/person/GenderEnum.java
@@ -1,0 +1,6 @@
+package br.com.solutis.locadora.model.entity.person;
+
+public enum GenderEnum {
+    MALE,
+    FEMALE,
+}
